+++ nest/common/decorators/core/__init__.py
@@ -1,0 +1,2 @@
+from .injectable_decorator import Injectable as Injectable
+
