+++ nest/core/__init__.py
@@ -1,1 +1,2 @@
 from .application_config import ApplicationConfig as ApplicationConfig
+from .nest_application import NestApplication as NestApplication
